requirements.txt
click==8.0.4
matplotlib==3.7.3
numpy==1.24.4
pandas==2.0.3
psutil==5.9.5
PyYAML==6.0.1
psycopg2-binary==2.9.9
